Add support for info messages during source stamp verification
apksig added support for the source stamp that allows an app
repository to add their own stamp on an APK to indicate the origin
of the APK. Currently the platform does not enforce the validity of
an APK's source stamp, so during verification apksig will only report
warnings if an APK's source stamp does not verify. However apksig
also offers support for a target that only performs source stamp
verification; this target will treat warnings as errors since all
stamp verification failures are reported as warnings. This
can cause some valid source stamps to fail such as when a new
attribute is added to the stamp that is unknown to previously
released verifiers. This commit adds support to stamp verification
to report certain events as info messages to prevent the events from
failing the verification; this info tag is currently applied to
unknown attributes and signature algorithms as these could be added
to stamps in the future and should not prevent existing verifiers
from properly verifying the stamp.

Bug: 231506876
Test: gradlew test
Change-Id: I94df013223a64c18d82518c82478325f516090df