Add APIs to disable source stamp timestamp attribute
The timestamp attribute was recently added to the source stamp to
allow verifies to determine the time at which the source stamp
was signed; however, since this attribute changes with each
signing, it can cause problems for reproducible builds. This commit
adds APIs to both apksig and a command line option to apksigner to
allow a caller to disable this attribute during a source stamp
signing.

Bug: 241036956
Test: gradlew test
Change-Id: If5586c07f2c8588faf6eedffccf64e664cb96bf8