Compare digests
We're comparing two set of digests. Each of them contains pairs: an
actual digets' bits and the digest algorithm that was used to compute
the digest.
We say that two sets match if they botho contain at least one common
digest algorithm and the actual digests are the same for every common
algorithm.

Bug: b/240942002

Change-Id: I91e08759e847d22adcb6a9d21cbbf3cad43e746f