Add API to change capabilities of node in lineage with signing cert
The apksig library supports modifying the capabilities of a previous
signer in the lineage, but this requires providing the previous
signer's full config (both certificate and private key). Since the
method only requires the previous signer's certificate, this commit
adds a new API to support modifying the capabilities of a previous
signer with only the signing certificate of the signer. The lineage
with capabilities is written to the signed data portion of the
current signer's signature block, so this API would not allow a
more relaxed lineage to be written to an existing APK without the
current signing key of that APK.

Bug: 246867486
Test: gradlew test
Change-Id: Ie623599f81718f6288f4c2acfab31da85ef29bef