Add support for SDK targeted signing configs
Android T added support for the V3.1 signature scheme which allows
an APK to target rotation for a specific SDK version and later. At
the time, apksig was updated to support this specific case of
rotation targeting, but generic SDK targeted signing configs could
not be specified. This commit adds APIs to apksig to allow these
generic signing configs to target SDK versions. This allows an
initial rotation for pre-T along with subsequent rotations for T+,
and it can also be used to specify different lineages with more
restrictive capabilities for later platform releases.

Bug: 218593443
Test: gradlew test
Change-Id: Ic6a3aa8511a8bd226baa142d687320b86f62303c