Handle OOM error when allocating space for uncompressed files
During APK V1 signature verification, all of the files in the APK
have to be uncompressed since jar signing computes and stores the
digest of the uncompressed file in the APK's manifest. During
the uncompress, the verifier will allocate a byte array for the
stated size of the file; if the verifier's JVM is not started with
sufficient heap space, this could result in an OutOfMemoryError.
This commit catches this error and adds it to an IOException to
allow the caller to properly handle it.

Bug: 266333259
Test: gradlew test
Change-Id: Iffc5b26dc31a1af4107f869442c1d296a922f9b3