Add zip64 support to apksig for LFH and CD records
Certain fields within an APK / zip file are subject to 32-bit
limitations; if the size of these fields exceed this, then a value
of 0xffffffff is written and the actual value is stored in a 64-bit
field in the Zip64 block within the extra field of the record. This
commit adds Zip64 support for these fields during APK signing /
signature verification; this will primarily be seen when an APK's
size is greater than 2GB, then local file header offsets will
be written in the Zip64 field. Note, this commit does not add
support for APKs > 4GB as this will change the structure of the
end of central directory and is not yet supported in the platform.

Bug: 302112430
Test: gradlew test
Change-Id: I29f980580dc4aa8744bf0cc9c719e574d9fa1f90