Add zip64 support to apksig for EoCD and CD records
When the number of entries in the central directory is greater
than 64K or when the offset of the central directory is beyond
4GB, the end of central directory (EoCD) record in the APK is
updated to also include a zip64 EoCD locator and a zip64 EoCD
record. This commit update the ZipUtils and EocdRecord classes
to handle both the parsing and writing of these zip64 sections.
Previous support was also added to handle zip64 fields in CD
records; this handled reading the zip64 fields from the extra
data in the CD record, but it did not include support to write
the appropriate extra data when recreating a record with deflate
compression. This commit updates the CentralDirectoryRecord
class to ensure this extra data can be written in a CD record
for the necessary zip64 fields.

Bug: 447631054
Flag: EXEMPT HOST_MODE
Test: gradlew test
Change-Id: Ie3ac6839e5dcc11ca0743ed7bb36a47fef4848b7